[Feat]#28 - WeatherDataProcessor 클래스 정의 - WeatherModel의 Weather 타입 중, description의 타입이 enum Description에서 String으로 수정, enum Description 미사용으로 인한 삭제 - forecast 배열에 데이터를 넣기 위한 단일책임의 클래스 WeatherDataProcessor 정의
